Export expressions and functions
