+++ src/schema/index.ts
@@ -1,2 +1,4 @@
 export { model, type Model } from '@/src/schema/model';
 export * from '@/src/schema/primitives';
+export { op } from '@/src/helpers/expressions';
+export * from '@/src/helpers/functions';
